+++ weblate/trans/models/component.py
@@ -542,7 +542,7 @@ class Component(models.Model, URLMixin, PathMixin):
         return AsyncResult(task_id)
 
     def progress_step(self, progress=None):
-        # No task (eg. eager mode)
+        # No task (for example eager mode)
         if not current_task or not current_task.request.id:
             return
         # Operate on linked component if needed